rename tests for better description
