+++ plenum/test/restart/test_restart_nodes_7_all.py
@@ -1,16 +1,12 @@
-import pytest
-from plenum.test.test_node import ensure_node_disconnected, checkNodesConnected
 from plenum.test import waits
 from plenum.test.node_request.helper import sdk_ensure_pool_functional
-from plenum.common.config_helper import PNodeConfigHelper
-from plenum.test.test_node import TestNode
-from plenum.test.restart.test_restart_nodes import get_group, restart_nodes
+from plenum.test.restart.helper import get_group, restart_nodes
 
 nodeCount = 7
 
 
-def test_restart_groups_wp(looper, txnPoolNodeSet, tconf, tdir,
-                           sdk_pool_handle, sdk_wallet_client, allPluginsPath):
+def test_restart_groups_7_of_7_wp_no_tm(looper, txnPoolNodeSet, tconf, tdir,
+                                        sdk_pool_handle, sdk_wallet_client, allPluginsPath):
     tm = tconf.ToleratePrimaryDisconnection + waits.expectedPoolElectionTimeout(len(txnPoolNodeSet))
 
     restart_group = get_group(txnPoolNodeSet, 7, include_primary=True)
